refactor: move constants.js to src/
